agregado dto y services de administradores.

+++ src/main/java/com/metalurgica1/metalurgica1/modelo/Persona.java
@@ -16,7 +16,7 @@ public abstract class Persona {
     private String contrasenia;
     private String nombre;
     private String telefono;
-    private int dni;
+    private Long dni;
 
     @Enumerated(EnumType.STRING)
     private EEtiquetaDeAcceso etiquetaDeAcceso;
